+++ lib/core/flame/components/hud/base_components/default_button.dart
@@ -1,4 +1,4 @@
-import 'package:defend_your_flame/core/flame/components/hud/abstract_components/text_button.dart';
+import 'package:defend_your_flame/core/flame/components/hud/base_components/text_button.dart';
 import 'package:defend_your_flame/core/flame/main_game.dart';
 import 'package:defend_your_flame/core/flame/managers/text_manager.dart';
 import 'package:flame/components.dart';
