+++ lib/core/flame/components/hud/buttons/enter_shop_button.dart
@@ -1,4 +1,4 @@
-import 'package:defend_your_flame/core/flame/components/hud/abstract_components/default_button.dart';
+import 'package:defend_your_flame/core/flame/components/hud/base_components/default_button.dart';
 import 'package:defend_your_flame/core/flame/components/hud/next_round_hud.dart';
 import 'package:defend_your_flame/core/flame/components/hud/next_round_internal/next_round_hud_state.dart';
 import 'package:defend_your_flame/core/flame/components/hud/next_round_internal/next_round_menu_hud.dart';
