+++ lib/core/flame/components/hud/buttons/go_back_button.dart
@@ -1,4 +1,4 @@
-import 'package:defend_your_flame/core/flame/components/hud/abstract_components/default_button.dart';
+import 'package:defend_your_flame/core/flame/components/hud/base_components/default_button.dart';
 import 'package:flutter/material.dart';
 
 class GoBackButton extends DefaultButton {
