+++ lib/core/flame/components/hud/buttons/restart_game_button.dart
@@ -1,4 +1,4 @@
-import 'package:defend_your_flame/core/flame/components/hud/abstract_components/default_button.dart';
+import 'package:defend_your_flame/core/flame/components/hud/base_components/default_button.dart';
 import 'package:defend_your_flame/core/flame/components/hud/game_over_hud.dart';
 import 'package:flame/components.dart';
