+++ lib/core/flame/components/hud/next_round_internal/next_round_menu_hud.dart
@@ -1,5 +1,5 @@
 import 'package:defend_your_flame/constants/theming_constants.dart';
-import 'package:defend_your_flame/core/flame/components/hud/abstract_components/basic_hud.dart';
+import 'package:defend_your_flame/core/flame/components/hud/base_components/basic_hud.dart';
 import 'package:defend_your_flame/core/flame/components/hud/buttons/enter_shop_button.dart';
 import 'package:defend_your_flame/core/flame/components/hud/buttons/next_round_button.dart';
 import 'package:defend_your_flame/core/flame/components/hud/next_round_hud.dart';
