+++ lib/core/flame/components/hud/text/round_text.dart
@@ -1,4 +1,4 @@
-import 'package:defend_your_flame/core/flame/components/hud/abstract_components/default_text.dart';
+import 'package:defend_your_flame/core/flame/components/hud/base_components/default_text.dart';
 import 'package:defend_your_flame/core/flame/managers/text_manager.dart';
 import 'package:defend_your_flame/core/flame/worlds/main_world.dart';
 import 'package:flame/components.dart';
