+++ lib/core/flame/components/hud/text/shop_title_text.dart
@@ -1,4 +1,4 @@
-import 'package:defend_your_flame/core/flame/components/hud/abstract_components/default_text.dart';
+import 'package:defend_your_flame/core/flame/components/hud/base_components/default_text.dart';
 
 class ShopTitleText extends DefaultText {
   @override
